Added new test for falsy values. Moved test to __tests__ folder
