+++ packages/get-safe-property-chain/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@empathy/get-safe-property-chain",
-  "version": "1.0.0-alpha.1",
+  "version": "1.0.0-alpha.2",
   "description": "A utility for safely retrieving nested values",
   "main": "dist/get-safe-property-chain.js",
   "types": "dist/get-safe-property-chain.d.ts",
